Add GitHub Action for Docker build and deploy
Introduce a new GitHub Action workflow in `.github/workflows/docker-deploy.yml` to automate Docker image build and deployment on push events to the main branch. The workflow consists of steps to check out the code, log in to Docker Hub, build and push the Docker image, update the Kubernetes configuration, and notify of success.